[add] Suggestion: pages と同じ名前のディレクトリが components ディレクトリに存在しなかったため kusa ディレクトリを追加した
